requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
picamera==1.13
pkg-resources==0.0.0
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
RPi.GPIO==0.7.0
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
